Fix #14
