+++ Tetris/StateGameMain.py
@@ -1,12 +1,13 @@
-from Tetris.State import *
 from Tetris.Tetromino import *
-from Tetris.StateMenuPause import PauseMenu
-from Tetris.StatePostGame import PostGame
+from Tetris.GameState import GameState
+from Tetris.StateMenuPause import StateMenuPause
+from Tetris.StateGamePost import StateGamePost
 
 import random
+import pygame
 
 
-class MainGame(GameState):
+class StateGameMain(GameState):
     def __init__(self, game, score: int = 0, multiplier: int = 1, grid: np.array = None, current_speed: int = None,
                  current_tetromino: dict = None, next_tetromino: dict = None, frame_counter: int = 0):
         super().__init__(game)
@@ -63,7 +64,7 @@ class MainGame(GameState):
                         pass
                     self.lock_movements = True  # On bloque les mouvements latéraux.
                 if event.type == pygame.KEYDOWN and event.key == pygame.K_ESCAPE:
-                    self.states_stack.append(PauseMenu(self.game))
+                    self.states_stack.append(StateMenuPause(self.game))
 
         self.frame_counter += 1
         if self.frame_counter >= self.current_speed:
@@ -75,7 +76,7 @@ class MainGame(GameState):
                 if not self.current_tetromino.try_move(0, 0, self.grid) and self.current_tetromino.y == 0:
                     # Game over
                     print("Game Over!")
-                    self.states_stack.append(PostGame(self.game, self.score))
+                    self.states_stack.append(StateGamePost(self.game, self.score))
                     return
 
                 self.score += self.score_for_new_shape_placed * self.multiplier
